pg_to_gcs_parquet/requirements.txt
psycopg2-binary>=2.9.9
polars>=0.20.0
pyarrow>=15.0.0
google-cloud-storage>=2.16.0
python-dotenv>=1.0.0
tqdm>=4.66.0
